+++ src/axislike.rs
@@ -30,6 +30,12 @@ pub struct SingleAxis {
     pub negative_low: f32,
     /// Whether to invert output values from this axis.
     pub inverted: bool,
+    /// How sensitive the axis is to input values.
+    ///
+    /// Since sensitivity is a multiplier, any value `>1.0` will increase sensitivity while any value `<1.0` will decrease sensitivity.
+    /// This value should always be strictly positive: a value of 0 will cause the axis to stop functioning,
+    /// while negative values will invert the direction.
+    pub sensitivity: f32,
     /// The target value for this input, used for input mocking.
     ///
     /// WARNING: this field is ignored for the sake of [`Eq`] and [`Hash`](std::hash::Hash)
@@ -45,6 +51,7 @@ impl SingleAxis {
             positive_low: threshold,
             negative_low: -threshold,
             inverted: false,
+            sensitivity: 1.0,
             value: None,
         }
     }
@@ -60,6 +67,7 @@ impl SingleAxis {
             positive_low: 0.0,
             negative_low: 0.0,
             inverted: false,
+            sensitivity: 1.0,
             value: Some(value),
         }
     }
@@ -72,6 +80,7 @@ impl SingleAxis {
             positive_low: 0.,
             negative_low: 0.,
             inverted: false,
+            sensitivity: 1.0,
             value: None,
         }
     }
@@ -84,6 +93,7 @@ impl SingleAxis {
             positive_low: 0.,
             negative_low: 0.,
             inverted: false,
+            sensitivity: 1.0,
             value: None,
         }
     }
@@ -96,6 +106,7 @@ impl SingleAxis {
             positive_low: 0.,
             negative_low: 0.,
             inverted: false,
+            sensitivity: 1.0,
             value: None,
         }
     }
@@ -108,6 +119,7 @@ impl SingleAxis {
             positive_low: 0.,
             negative_low: 0.,
             inverted: false,
+            sensitivity: 1.0,
             value: None,
         }
     }
@@ -121,6 +133,7 @@ impl SingleAxis {
             negative_low: threshold,
             positive_low: f32::MAX,
             inverted: false,
+            sensitivity: 1.0,
             value: None,
         }
     }
@@ -134,6 +147,7 @@ impl SingleAxis {
             negative_low: f32::MIN,
             positive_low: threshold,
             inverted: false,
+            sensitivity: 1.0,
             value: None,
         }
     }
@@ -143,6 +157,13 @@ impl SingleAxis {
     pub fn with_deadzone(mut self, deadzone: f32) -> SingleAxis {
         self.negative_low = -deadzone;
         self.positive_low = deadzone;
+        self
+    }
+
+    /// Returns this [`SingleAxis`] with the sensitivity set to the specified value
+    #[must_use]
+    pub fn with_sensitivity(mut self, sensitivity: f32) -> SingleAxis {
+        self.sensitivity = sensitivity;
         self
     }
 
@@ -159,6 +180,7 @@ impl PartialEq for SingleAxis {
         self.axis_type == other.axis_type
             && FloatOrd(self.positive_low) == FloatOrd(other.positive_low)
             && FloatOrd(self.negative_low) == FloatOrd(other.negative_low)
+            && FloatOrd(self.sensitivity) == FloatOrd(other.sensitivity)
     }
 }
 impl Eq for SingleAxis {}
@@ -167,6 +189,7 @@ impl std::hash::Hash for SingleAxis {
         self.axis_type.hash(state);
         FloatOrd(self.positive_low).hash(state);
         FloatOrd(self.negative_low).hash(state);
+        FloatOrd(self.sensitivity).hash(state);
     }
 }
 
@@ -278,6 +301,14 @@ impl DualAxis {
     #[must_use]
     pub fn with_deadzone(mut self, deadzone: DeadZoneShape) -> DualAxis {
         self.deadzone = deadzone;
+        self
+    }
+
+    /// Returns this [`DualAxis`] with the sensitivity set to the specified values
+    #[must_use]
+    pub fn with_sensitivity(mut self, x_sensitivity: f32, y_sensitivity: f32) -> DualAxis {
+        self.x.sensitivity = x_sensitivity;
+        self.y.sensitivity = y_sensitivity;
         self
     }