+++ tests/gamepad_axis.rs
@@ -79,6 +79,7 @@ fn game_pad_single_axis_mocking() {
         value: Some(-1.),
         positive_low: 0.0,
         negative_low: 0.0,
+        sensitivity: 1.0,
         inverted: false,
     };
 
@@ -99,6 +100,7 @@ fn game_pad_dual_axis_mocking() {
             value: Some(1.),
             positive_low: 0.0,
             negative_low: 0.0,
+            sensitivity: 1.0,
             inverted: false,
         },
         y: SingleAxis {
@@ -106,6 +108,7 @@ fn game_pad_dual_axis_mocking() {
             value: Some(0.),
             positive_low: 0.0,
             negative_low: 0.0,
+            sensitivity: 1.0,
             inverted: false,
         },
         deadzone: DualAxis::DEFAULT_DEADZONE_SHAPE,
@@ -137,6 +140,7 @@ fn game_pad_single_axis() {
         positive_low: 0.0,
         negative_low: 0.0,
         inverted: false,
+        sensitivity: 1.0,
     };
     app.send_input(input);
     app.update();
@@ -150,6 +154,7 @@ fn game_pad_single_axis() {
         positive_low: 0.0,
         negative_low: 0.0,
         inverted: false,
+        sensitivity: 1.0,
     };
     app.send_input(input);
     app.update();
@@ -163,6 +168,7 @@ fn game_pad_single_axis() {
         positive_low: 0.0,
         negative_low: 0.0,
         inverted: false,
+        sensitivity: 1.0,
     };
     app.send_input(input);
     app.update();
@@ -176,6 +182,7 @@ fn game_pad_single_axis() {
         positive_low: 0.0,
         negative_low: 0.0,
         inverted: false,
+        sensitivity: 1.0,
     };
     app.send_input(input);
     app.update();
@@ -190,6 +197,7 @@ fn game_pad_single_axis() {
         positive_low: 0.1,
         negative_low: 0.1,
         inverted: false,
+        sensitivity: 1.0,
     };
     app.send_input(input);
     app.update();
@@ -203,11 +211,88 @@ fn game_pad_single_axis() {
         positive_low: 0.0,
         negative_low: 0.0,
         inverted: false,
+        sensitivity: 1.0,
     };
     app.send_input(input);
     app.update();
     let action_state = app.world.resource::<ActionState<AxislikeTestAction>>();
     assert!(!action_state.pressed(AxislikeTestAction::Y));
+}
+
+#[test]
+fn game_pad_single_axis_inverted() {
+    let mut app = test_app();
+    app.insert_resource(InputMap::new([
+        (
+            SingleAxis::symmetric(GamepadAxisType::LeftStickX, 0.1).inverted(),
+            AxislikeTestAction::X,
+        ),
+        (
+            SingleAxis::symmetric(GamepadAxisType::LeftStickY, 0.1).inverted(),
+            AxislikeTestAction::Y,
+        ),
+    ]));
+
+    // +X
+    let input = SingleAxis {
+        axis_type: AxisType::Gamepad(GamepadAxisType::LeftStickX),
+        value: Some(1.),
+        positive_low: 0.0,
+        negative_low: 0.0,
+        inverted: true,
+        sensitivity: -1.0,
+    }
+    .inverted();
+    app.send_input(input);
+    app.update();
+    let action_state = app.world.resource::<ActionState<AxislikeTestAction>>();
+    assert!(action_state.pressed(AxislikeTestAction::X));
+    assert!(action_state.value(AxislikeTestAction::X) == -1.0);
+
+    // -X
+    let input = SingleAxis {
+        axis_type: AxisType::Gamepad(GamepadAxisType::LeftStickX),
+        value: Some(-1.),
+        positive_low: 0.0,
+        negative_low: 0.0,
+        inverted: true,
+        sensitivity: -1.0,
+    };
+    app.send_input(input);
+    app.update();
+    let action_state = app.world.resource::<ActionState<AxislikeTestAction>>();
+    assert!(action_state.pressed(AxislikeTestAction::X));
+    assert!(action_state.value(AxislikeTestAction::X) == 1.0);
+
+    // +Y
+    let input = SingleAxis {
+        axis_type: AxisType::Gamepad(GamepadAxisType::LeftStickY),
+        value: Some(1.),
+        positive_low: 0.0,
+        negative_low: 0.0,
+        inverted: true,
+        sensitivity: -1.0,
+    };
+    app.send_input(input);
+    app.update();
+    let action_state = app.world.resource::<ActionState<AxislikeTestAction>>();
+    assert!(action_state.pressed(AxislikeTestAction::Y));
+    assert!(action_state.value(AxislikeTestAction::Y) == -1.0);
+
+    // -Y
+    let input = SingleAxis {
+        axis_type: AxisType::Gamepad(GamepadAxisType::LeftStickY),
+        value: Some(-1.),
+        positive_low: 0.0,
+        negative_low: 0.0,
+        inverted: true,
+        sensitivity: -1.0,
+    };
+    app.send_input(input);
+    app.update();
+    let action_state = app.world.resource::<ActionState<AxislikeTestAction>>();
+    assert!(action_state.pressed(AxislikeTestAction::Y));
+    assert!(action_state.value(AxislikeTestAction::Y) == 1.0);
 }
 
 #[test]